Now ratio an be stored as float for better contorl

+++ flomo/config.py
@@ -98,7 +98,7 @@ class Config:
                     tag, name, ratio = value.split(" ")
                     data[key]["tag"] = tag
                     data[key]["name"] = name
-                    data[key]['ratio'] = int(ratio)
+                    data[key]['ratio'] = float(ratio)
                 elif key == TAG_COLORS:
                     key_, value_ = value.split(" ")
                     data[key][key_] = value_
